MDEV-34413 Index Condition Pushdown for reverse-ordered scans
Corrects ICP accounting during the handler callback by including
range misses as part of the attempted ICP counter.  Adds a new
counter, 'ICP misses', which is incremented when an ICP attempt
is made at the handler but does not result in a match.  Updates
existing tests where stats need to be recomputed based on the
changes to the attempted ICP counter.

Adds tests which show that ICP was not enabled for reverse-ordered scans
prior to this mdev.  A later commit for this same mdev records again
these same tests, showing that ICP for reverse-ordered scans is
enabled and working.