MDEV-10164: Add support for TRIGGERS that fire on multiple events
Added capability to create a trigger associated with several trigger events.
For this goal, the syntax of the CREATE TRIGGER statement was extended
to support the syntax structure { event [ OR ... ] } for the `trigger_event`
clause.

Support of several trigger events for the same trigger requires that
the data members `next` and `actoion_order` of the Trigger class
be converted to arrays to store relating information per trigger event base.